requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==2.6.0
Flask-Script==2.0.5
psycopg2==2.9.1
pandas==1.2.5
python-dotenv
flasgger==0.9.5
pytest==6.2.4
marshmallow==3.15.0
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.0
werkzeug==2.0.3
